requirements.txt
Django>=5.0.0
Pillow>=10.0.0
django-tinymce>=3.7.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
